luxemart_base/requirements.txt
Django>=4.2,<5.0
numpy>=1.24
scikit-learn>=1.3
pandas>=2.0
Pillow>=10.0
six>=1.16
